setup: organize
added folders for agents and move ocnfig file into core
